requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
dnspython==2.2.1
docutils==0.19
et-xmlfile==1.1.0
idna==3.4
Kivy==2.1.0
Kivy-Garden==0.1.5
kivymd==1.1.1
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
Pillow==9.2.0
Pygments==2.13.0
pymongo==4.3.2
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
urllib3==1.26.12
